requirements.txt
torch
transformers
datasets
sounddevice
pyannote.audio
pydub
Jinja2
PySimpleGUI
pyaudio
numpy